added prisma schema, moved components to src
