--- lib/apollo/ApolloClient.ts
@@ -1,113 +1,0 @@
-import { useMemo } from "react";
-import {
-  ApolloClient,
-  HttpLink,
-  InMemoryCache,
-  from,
-  NormalizedCacheObject,
-} from "@apollo/client";
-import { onError } from "@apollo/client/link/error";
-import merge from "deepmerge";
-import isEqual from "lodash/isEqual";
-import { ApolloCacheConfig } from "./ApolloCacheConfig";
-
-export const APOLLO_STATE_PROP_NAME = "__APOLLO_STATE__";
-
-let apolloClient: ApolloClient<NormalizedCacheObject>;
-
-const errorLink = onError(({ graphQLErrors, networkError }) => {
-  if (graphQLErrors)
-    graphQLErrors.forEach(({ message, locations, path }) =>
-      console.log(
-        `[GraphQL error]: Message: ${message}, Location: ${JSON.stringify(
-          locations,
-          null,
-          2
-        )}, Path: ${path}`
-      )
-    );
-  if (networkError) console.log(`[Network error]: ${networkError}`);
-});
-
-function createApolloClient(Cookie?: string) {
-  return new ApolloClient({
-    ssrMode: typeof window === "undefined",
-    link: from([
-      errorLink,
-      new HttpLink({
-        uri:
-          process.env.NODE_ENV == "production"
-            ? "https://dorten.vercel.app/api/graphql"
-            : "http://localhost:3000/api/graphql",
-        // Server URL (must be absolute)
-        credentials: "include", // Additional fetch() options like `credentials` or `headers`
-        headers: {
-          Cookie,
-        },
-      }),
-    ]),
-    cache: new InMemoryCache(ApolloCacheConfig),
-  });
-}
-const CookieObjToString = (
-  Cookie?: Partial<{
-    [key: string]: string;
-  }>
-) => {
-  if (!Cookie) return "";
-  const cookieKey = Object.keys(Cookie)[0];
-  return `${cookieKey}=${Cookie[cookieKey]}`;
-};
-export function initializeApollo(
-  initialState: NormalizedCacheObject | null = null,
-  cookie?: Partial<{
-    [key: string]: string;
-  }>
-) {
-  const Cookie = CookieObjToString(cookie);
-  const _apolloClient = apolloClient ?? createApolloClient(Cookie);
-
-  // If your page has Next.js data fetching methods that use Apollo Client, the initial state
-  // gets hydrated here
-  if (initialState) {
-    // Get existing cache, loaded during client side data fetching
-    const existingCache = _apolloClient.extract();
-
-    // Merge the initialState from getStaticProps/getServerSideProps in the existing cache
-    const data = merge(existingCache, initialState, {
-      // combine arrays using object equality (like in sets)
-      arrayMerge: (destinationArray, sourceArray) => [
-        ...sourceArray,
-        ...destinationArray.filter((d) =>
-          sourceArray.every((s) => !isEqual(d, s))
-        ),
-      ],
-    });
-
-    // Restore the cache with the merged data
-    _apolloClient.cache.restore(data);
-  }
-  // For SSG and SSR always create a new Apollo Client
-  if (typeof window === "undefined") return _apolloClient;
-  // Create the Apollo Client once in the client
-  if (!apolloClient) apolloClient = _apolloClient;
-
-  return _apolloClient;
-}
-
-export function addApolloState(
-  client: ApolloClient<NormalizedCacheObject>,
-  pageProps: any
-) {
-  if (pageProps?.props) {
-    pageProps.props[APOLLO_STATE_PROP_NAME] = client.cache.extract();
-  }
-
-  return pageProps;
-}
-
-export function useApollo(pageProps: any) {
-  const state = pageProps[APOLLO_STATE_PROP_NAME];
-  const store = useMemo(() => initializeApollo(state), [state]);
-  return store;
-}
