+++ package.json
@@ -59,6 +59,7 @@
     "postcss": "^8.4.16",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.13",
+    "prisma": "^4.8.0",
     "tailwindcss": "^3.1.8",
     "ts-node": "^10.9.1",
     "typescript": "4.7.4"
