+++ prisma/schema.prisma
@@ -1,0 +1,75 @@
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABASE_URL")
+}
+
+model account {
+  provider   account_provider_enum
+  account_id String                @id @db.VarChar
+  created_at DateTime              @default(now()) @db.Timestamp(6)
+  user       user?                 @relation(fields: [account_id], references: [user_id], onDelete: NoAction, onUpdate: NoAction)
+}
+
+model comment {
+  comment_id    String   @id @default(cuid()) @db.VarChar
+  text          String   @db.VarChar
+  post_id       String   @db.VarChar
+  creator_id    String   @db.VarChar
+  created_at    DateTime @default(now()) @db.Timestamp(6)
+  user          user?    @relation(fields: [creator_id], references: [user_id], onDelete: NoAction, onUpdate: NoAction)
+  post          post?    @relation(fields: [post_id], references: [post_id], onDelete: NoAction, onUpdate: NoAction)
+}
+
+model like {
+  value      Int
+  userId     String  @db.VarChar
+  postId     String  @db.VarChar
+  user       user?   @relation(fields: [userId], references: [user_id], onDelete: NoAction, onUpdate: NoAction)
+  post       post?   @relation(fields: [postId], references: [post_id], onDelete: Cascade, onUpdate: NoAction)
+
+  @@id([userId, postId])
+}
+
+model post {
+  post_id       String    @id @default(cuid()) @db.VarChar
+  title         String    @db.VarChar
+  description   String?   @db.VarChar
+  picture       String?   @db.VarChar
+  creator_id    String    @db.VarChar
+  points        Int       @default(0)
+  created_at    DateTime  @default(now()) @db.Timestamp(6)
+  comment       comment[]
+  like          like[]
+  user          user?     @relation(fields: [creator_id], references: [user_id], onDelete: NoAction, onUpdate: NoAction)
+}
+
+model user {
+  user_id        String           @id @default(cuid()) @db.VarChar
+  username       String           @unique @db.VarChar
+  email          String           @unique @db.VarChar
+  email_verified Boolean          @default(false)
+  password       String?          @db.VarChar
+  picture        String?          @db.VarChar
+  age            Int?
+  gender         user_gender_enum @default(None)
+  created_at     DateTime         @default(now()) @db.Timestamp(6)
+  account        account[]
+  comment        comment[]
+  like           like[]
+  post           post[]
+}
+
+enum account_provider_enum {
+  GOOGLE
+  GITHUB
+}
+
+enum user_gender_enum {
+  Male
+  Female
+  None
+}
