+++ src/components/General/PostCard/LCS.tsx
