+++ src/components/Layout.tsx
@@ -1,7 +1,7 @@
 import { Router } from "next/router";
 import { useState, useEffect } from "react";
 import { BarLoader } from "react-spinners";
-import { useColorMode } from "../Hooks/useColorMode";
+import { useColorMode } from "../src/Hooks/useColorMode";
 import { theme } from "../tailwind.config";
 import { Copyright } from "./Copyright";
 import NavBar from "./NavBar/NavBar";
